# New: code path analysis.