Fixes #3530 

- [x] Decide API (new events style v.s. `context.getCodePath()` style)
- [x] Add JSDoc comments.
- [x] Add more tests.
  - [x] logical expressions.
  - [x] if statements.
  - [x] switch statements.
  - [x] try-catch-finally statements.
  - [x] while statements.
  - [x] do-while statements.
  - [x] for statements.
  - [x] for-in/of statements.
  - [x] functions.
  - [x] default parameters.
  - [x] default values of destructuring assignments
- [x] Add documents.
- [x] Remake some rules with the code path analysis.
  - [x] `no-unreachable`
  - [x] `no-fallthrough`
  - [x] `constructor-super`
  - [x] `no-this-before-super`
  - [x] `consistent-return`