@mysticatea 2015-08-28
Currently, I'm adding tests and fixing comments around `switch`...