@mysticatea 2015-08-30
I'm adding tests for `try-catch-finally` currently.

I'm considering paths in `try` blocks.

```js
class A {
    constructor(a) {
        try {
            super(a);
        } catch (err) {
        }

        this.foo();
    }
}
```

This code does not throw any error before calling `super(a);`.
My plan is creating two paths from `try` block to `catch` block.

- The first throwable node (e.g. `Identifier` as a reference)
- The last segment in `try` block.
- (and at `throw` statements)

Because those are at opposite poles. I guess most rules is good enough to use the two paths.
But in this case, `a` is `Identifier` as a reference, but not throwable. 
In order to detect whether the Identifier is throwable or not, resolving names is needed.