@nzakas 2015-08-31
As far as I know, you can't use `this` if `super()` throws.

Regarding whether to fork or not, is there any existing code path analysis tool that might be useful to look at? For instance, maybe Flow does this and we could figure out the right way to handle it based on that?

@jeffmo are you all doing this type of code path analysis? Any insights?