@jeffmo 2015-08-31
I'm pretty sure we don't track partial initialization state right now except in a few very trivial circumstances. It's a TODO for us, but probably not in the short term.

This is pretty neat, though. I imagine when we do get to this we'll model it like any other variable initialization and ask the user to consider "potentially uninitialized" for `this` as well. `super()` isn't a normal "call" by pretty much any definition...so I wouldn't feel too bad about special-casing it as 'throwable'.

P.S. Just tossing in the spec text confirming what you said regarding `this` not being initialized if `super()` throws ("SuperCall : super Arguments" - Steps 8-10):

http://www.ecma-international.org/ecma-262/6.0/index.html#sec-super-keyword-runtime-semantics-evaluation