@mysticatea 2015-09-02
@nzakas @jeffmo Thank you very much.

escope have completed scope analysis in this time, so I guessed we can detect whether the Identifier is throwable or not. But escope does not have `O(1)` getter for variables...

So I will do not detecting that to start. I will think below as throwable:

- All `Identifier` as a reference.
- All `MemberExpression` (can throw a `null` access error, or might be a getter)
- All `CallExpression` (includes `super()`)
- All `NewExpression`