@mysticatea 2015-09-25
Yes, I'm working on this. I'm sorry for slowly.
I have time in weekends mainly (I have late on my job now) and I'm setting higher priority to fix bugs.