@mysticatea 2015-10-12
Done, ready for review!

I have run the rewritten rules in some repositories (eslint, escope, espree, npm, and mine), then I didn't find wrong results.

This PR contains:

- Code Path Analyzer's code (`lib/code-path-analysis/*`).
- Code Path Analyzer's tests (`tests/lib/code-path-analysis/*` and `tests/fixtures/code-path-analysis/*`).
- 5 rules are rewritten with code paths.
- Some code modifications that rewritten rules have found.
- Polyfill for `Function.prototype.bind` of PhantomJS.
- A document (`docs/developer-guide/working-with-rules.md`)

----

Main files:

- `code-path.js`
- `code-path-analyzer.js`
- `code-path-segment.js`
- `code-path-state.js`
- `debug-helpers.js`
- `fork-context.js`
- `id-generator.js`

Most process is in `code-path-analyzer.js` and `code-path-state.js`.
`code-path-analyzer.js` traverses AST and gives information to `code-path-state.js`.
`code-path-state.js` manages the state of a code path with given information from `code-path-analyzer.js`.
`code-path.js` and `code-path-segment.js` are public interface. 
`debug-helpers.js`, `fork-context.js`, and `id-generator.js` are internal helpers.