@gyandeeps 2015-10-12
Before we can start reviewing the code, can make sure your branch is rebased with the latest master.