@gyandeeps 2015-10-12
Quick question:
* What happens if not all code paths return? Like the example below
```js
function test(x){
  if(x === 1){
    console.log("do nothing");
  }
  else{
    return true;
  }
}
```

* Similar question but for switch cases where if I dont define a default case?
* Can you also share the perf run for before and after this code?

Side note: To be completely honest here (:smile:) I dont think I can completely digest this whole logic in one go. I have to run through and play around with code to grasp how all this works.