@ilyavolodin 2015-10-12
I completely agree with @gyandeeps this is just too much code to go through in a single code review. @mysticatea any ideas if we can somehow split this up into multiple PRs? For example we should be able to easily separate rule updates into it's own PR. Anything else that can be pulled out?
This is one of the most complicated features in the whole code-base. I think it deserves thorough code review, and I'm very excited about possibilities that this opens up. But going through 130 files is just too much.