@nzakas 2015-10-12
This is an amazing amount of work, thanks @mysticatea! 

I agree with @ilyavolodin, I think we should split this into a some smaller pull requests to make it easier for people to review.

I'm also a little bit concerned about the note regarding `getCodePath()` returning an incomplete `CodePath` object, as that seems like it could be confusing to people and potentially cause errors.

We're also going to have to do a lot of documentation around how this works. I'm not sure the difference between a `CodePath` and `CodePathSegment`, is there any place I can go to read more about this approach?