@mysticatea 2015-10-12
@gyandeeps 

> What happens if not all code paths return? Like the example below

If all paths are not returned, it addresses that the final segment is returned.

![image](https://cloud.githubusercontent.com/assets/1937871/10441589/b47b583a-7185-11e5-883b-8b152576d5d7.png)

> Similar question but for switch cases where if I dont define a default case?

There are test cases here: `tests\fixtures\code-path-analysis\switch--cases-*`
It goes the next code after walking all case tests.

> Can you also share the perf run for before and after this code?

BEFORE:

```
CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  1972.3731480000001ms
Performance Run #2:  1957.460232ms
Performance Run #3:  1983.6695949999998ms
Performance Run #4:  1969.877751ms
Performance Run #5:  1982.227052ms
Performance budget ok:  1972.3731480000001ms (limit: 2211.0849056603774ms)


Load performance Run #1:  234.612575ms
Load performance Run #2:  233.000996ms
Load performance Run #3:  232.582933ms
Load performance Run #4:  230.61668ms
Load performance Run #5:  232.313381ms

Load Performance median :  232.582933ms
```

AFTER:

```
CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  2168.29696ms
Performance Run #2:  2165.601437ms
Performance Run #3:  2187.562861ms
Performance Run #4:  2165.585137ms
Performance Run #5:  2159.003894ms
Performance budget ok:  2165.601437ms (limit: 2211.0849056603774ms)


Load performance Run #1:  242.883576ms
Load performance Run #2:  239.578617ms
Load performance Run #3:  238.709589ms
Load performance Run #4:  240.879894ms
Load performance Run #5:  238.128226ms

Load Performance median :  239.578617ms
```

About 10% slower :cry: 

----

There are many test cases in `tests\fixtures\code-path-analysis`.
Each test case has js code and DOT code to visualize.
If we copy and paste DOT code to http://www.webgraphviz.com/ , we can look at a code path as dygraph.

Also, debug mode (`DEBUG="eslint:code-path"`) outputs DOT codes of each code path.