@mysticatea 2015-10-31
I modified some functions that a profiler says it's slow.

- `RuleContext` constructor (400ms -> 1ms)
- `getNodeIndent` function (160ms -> 8ms)
- etc...

Then I modified `npm run perf` (`eslint --no-ignore ./tests/performance/jshint.js` --> `eslint --no-ignore lib`) and ran.

As results:

**Original**

```
CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  3337.679209ms
Performance Run #2:  3377.260864ms
Performance Run #3:  3337.500818ms
Performance Run #4:  3295.8551429999998ms
Performance Run #5:  3313.856627ms
Performance budget exceeded: 3337.500818ms (limit: 2211.0849056603774ms)
```

**Added Code Path Analysis**

```
CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  3832.92408ms
Performance Run #2:  3800.501909ms
Performance Run #3:  3816.432429ms
Performance Run #4:  3836.903315ms
Performance Run #5:  3832.869748ms
Performance budget exceeded: 3832.869748ms (limit: 2211.0849056603774ms)
```

500ms slowly... I'm sad....

**Added Code Path Analysis + Improved Performance**

```
CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  3238.026604ms
Performance Run #2:  3136.37427ms
Performance Run #3:  3163.708829ms
Performance Run #4:  3153.404716ms
Performance Run #5:  3131.907554ms
Performance budget exceeded: 3153.404716ms (limit: 2211.0849056603774ms)
```

About 150ms faster then original.

-----

-----

On the other hand, performance for one file linting is still slower.